Move delete workflow to api package
